+++ spyder/widgets/tabs.py
@@ -16,11 +16,14 @@ import os.path as osp
 
 # Third party imports
 import qstylizer.style
+from qtpy import QT6
 from qtpy.QtCore import QEvent, QPoint, Qt, Signal, Slot, QSize
 from qtpy.QtGui import QFontMetrics
 from qtpy.QtWidgets import (
+    QCommonStyle,
     QHBoxLayout,
     QLineEdit,
+    QStyle,
     QTabBar,
     QTabWidget,
     QToolButton,
@@ -28,6 +31,7 @@ from qtpy.QtWidgets import (
 )
 
 # Local imports
+from spyder.api.config.mixins import SpyderConfigurationAccessor
 from spyder.api.shortcuts import SpyderShortcutsMixin
 from spyder.api.translations import _
 from spyder.api.widgets.menus import SpyderMenu
@@ -36,12 +40,7 @@ from spyder.utils.misc import get_common_path
 from spyder.utils.palette import SpyderPalette
 from spyder.utils.qthelpers import (add_actions, create_action,
                                     create_toolbutton)
-from spyder.utils.stylesheet import (
-    CLOSE_BUTTON_TABBAR_STYLE,
-    MAC,
-    PANES_TABBAR_STYLESHEET,
-    WIN
-)
+from spyder.utils.stylesheet import MAC, PANES_TABBAR_STYLESHEET, WIN
 
 
 class EditTabNamePopup(QLineEdit):
@@ -258,6 +257,39 @@ class CloseTabButton(QToolButton):
         self.css.setValues(backgroundColor=f'{background_color}')
         self.setStyleSheet(self.css.toString())
 
+
+class CloseButtonTabBarStyle(QCommonStyle, SpyderConfigurationAccessor):
+    """
+    Style for QTabBar instances which should pull
+    QStyle.SH_TabBar_CloseButtonPosition from a CONF entry.
+    """
+    def __init__(self, *args, **kwargs):
+        super().__init__(*args, **kwargs)
+        position = self.get_conf(
+            "tab_close_position",
+            section="main"
+        )
+        if position == "left":
+            self.close_btn_side = QTabBar.ButtonPosition.LeftSide
+        elif position == "right":
+            self.close_btn_side = QTabBar.ButtonPosition.RightSide
+        else:
+            if MAC:
+                self.close_btn_side = QTabBar.ButtonPosition.LeftSide
+            else:
+                self.close_btn_side = QTabBar.ButtonPosition.RightSide
+
+    def styleHint(self, hint, options=None, widget=None, returnData=None):
+        if hint == QStyle.SH_TabBar_CloseButtonPosition:
+            if QT6:  # PySide6/PyQt6
+                return self.close_btn_side.value
+            else:  # PySide2/PyQt5
+                return int(self.close_btn_side)
+        else:
+            return super().styleHint(hint, options, widget, returnData)
+
+
+CLOSE_BUTTON_TABBAR_STYLE = CloseButtonTabBarStyle()
 
 class TabBar(QTabBar):
     """Tabs base class with drag and drop support"""
